+++ src/context/SøknadContext.tsx
@@ -20,6 +20,8 @@ import { IMellomlagretOvergangsstønad } from '../models/søknad/mellomlagretSø
 import Environment from '../Environment';
 import { useIntl } from 'react-intl';
 import { MellomlagredeStønadstyper } from '../models/søknad/stønadstyper';
+import { IBarn } from '../models/steg/barn';
+import { oppdaterObjektIListe } from '../utils/array';
 
 // -----------  CONTEXT  -----------
 const initialState: ISøknad = {
@@ -153,6 +155,23 @@ const [SøknadProvider, useSøknad] = createUseContext(() => {
     });
   };
 
+  const oppdaterBarnISoknaden = (
+    oppdatertBarn: IBarn,
+    indeksBarnSomErHuket: number
+  ) => {
+    settSøknad((prevSøknad) => ({
+      ...prevSøknad,
+      person: {
+        ...prevSøknad.person,
+        barn: oppdaterObjektIListe<IBarn>(
+          søknad.person.barn,
+          oppdatertBarn,
+          indeksBarnSomErHuket
+        ),
+      },
+    }));
+  };
+
   return {
     søknad,
     settSøknad,
@@ -163,6 +182,7 @@ const [SøknadProvider, useSøknad] = createUseContext(() => {
     mellomlagreOvergangsstønad,
     brukMellomlagretOvergangsstønad,
     nullstillMellomlagretOvergangsstønad,
+    oppdaterBarnISoknaden,
   };
 });
